+++ src/main/java/com/Twitter/org/controllers/LikesController.java
@@ -2,35 +2,59 @@ package com.Twitter.org.controllers;
 
 import com.Twitter.org.Models.Response;
 import com.Twitter.org.Models.Tweets.Tweets;
-import com.Twitter.org.Models.dto.TweetsDto;
-import com.Twitter.org.Models.dto.UserDto;
-import com.Twitter.org.mappers.Impl.TweetsMapper;
-import com.Twitter.org.services.LikesService;
+import com.Twitter.org.Models.dto.TweetsDto.TweetsDetailsDto;
+import com.Twitter.org.Models.dto.UserDto.UserResponseDto;
+import com.Twitter.org.mappers.Impl.UserMapper.UserResponseMapper;
+import com.Twitter.org.mappers.Mapper;
+import com.Twitter.org.services.Likes.LikesService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@Tag(name = "Likes", description = "Likes operations")
 @RestController
 public class LikesController {
 
     private final LikesService likesService;
-    private final TweetsMapper tweetsMapper;
+    private final UserResponseMapper userResponseMapper;
+    private final Mapper<Tweets, TweetsDetailsDto> TweetsDetailsDtoMapper;
 
 
-    public LikesController(LikesService likesService, TweetsMapper tweetsMapper) {
+    public LikesController(LikesService likesService, UserResponseMapper userResponseMapper, Mapper<Tweets, TweetsDetailsDto> tweetsDetailsDtoMapper) {
         this.likesService = likesService;
-        this.tweetsMapper = tweetsMapper;
+        this.userResponseMapper = userResponseMapper;
+        TweetsDetailsDtoMapper = tweetsDetailsDtoMapper;
+    }
+
+    private static ResponseEntity<?> GetUserDetailsResponse(String userName) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String authUserName = userDetails.getUsername();
+        boolean isAdmin = userDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+        // If the user is not authenticated or not an admin
+        if (!authUserName.equals(userName) && !isAdmin) {
+            return ResponseEntity.status(401).body("Unauthorized or Unauthenticated");
+        }
+        return null;
     }
 
 
     // User likes a tweet
+    @Operation(summary = "Like a tweet")
     @PostMapping("/{username}/tweets/{tweetId}/likes")
-    public ResponseEntity<Response> likeTweet(@PathVariable String username, @PathVariable int tweetId) {
+    public ResponseEntity<?> likeTweet(@PathVariable String username, @PathVariable int tweetId) {
+        final ResponseEntity<?> Unauthorized_or_Unauthenticated = GetUserDetailsResponse(username);
+        if (Unauthorized_or_Unauthenticated != null) return Unauthorized_or_Unauthenticated;
+
         Response response = likesService.addLike(username, tweetId);
         if (response.isSuccess()) {
             // Map the Tweets object to a TweetsDto object
-            TweetsDto tweetDto = tweetsMapper.mapTo((Tweets) response.getData());
+            TweetsDetailsDto tweetDto = TweetsDetailsDtoMapper.mapTo((Tweets) response.getData());
             // Set the TweetsDto object to the response data
             response.setData(tweetDto);
             return ResponseEntity.ok(response);
@@ -39,11 +63,16 @@ public class LikesController {
     }
 
     // User unlikes a tweet
+    @Operation(summary = "Delete a like on a tweet")
     @DeleteMapping("/{username}/tweets/{tweetId}/likes")
-    public ResponseEntity<Response> unlikeTweet(@PathVariable String username, @PathVariable int tweetId) {
+    public ResponseEntity<?> unlikeTweet(@PathVariable String username, @PathVariable int tweetId) {
+        final ResponseEntity<?> Unauthorized_or_Unauthenticated = GetUserDetailsResponse(username);
+        if (Unauthorized_or_Unauthenticated != null) return Unauthorized_or_Unauthenticated;
+
+
         Response response = likesService.removeLike(username, tweetId);
         if (response.isSuccess()) {
-            TweetsDto tweetDto = tweetsMapper.mapTo((Tweets) response.getData());
+            TweetsDetailsDto tweetDto = TweetsDetailsDtoMapper.mapTo((Tweets) response.getData());
             response.setData(tweetDto);
             return ResponseEntity.ok(response);
         }
@@ -51,35 +80,40 @@ public class LikesController {
     }
 
     // Check if a user has liked a tweet
+    @Operation(summary = "Check user liked a tweet")
     @GetMapping("/{username}/tweets/{tweetId}/likes")
-    public boolean isLiked(@PathVariable String username, @PathVariable int tweetId) {
-        return likesService.isLiked(username, tweetId);
+    public ResponseEntity<?> isLiked(@PathVariable String username, @PathVariable int tweetId) {
+        return ResponseEntity.ok(likesService.isLiked(username, tweetId));
     }
 
     // Count the number of likes for a tweet
+    @Operation(summary = "Count likes of a tweet")
     @GetMapping("/tweets/{tweetId}/likes/count")
-    public int countLikes(@PathVariable int tweetId) {
-        return likesService.countLikes(tweetId);
+    public ResponseEntity<?> countLikes(@PathVariable int tweetId) {
+        return ResponseEntity.ok(likesService.countLikes(tweetId));
     }
 
     // Get all tweets liked by a user
+    @Operation(summary = "Get all liked tweet by a user")
     @GetMapping("/{username}/likes/tweets")
-    public List<TweetsDto> getLikedTweets(@PathVariable String username) {
+    public ResponseEntity<?> getLikedTweets(@PathVariable String username) {
         List<Tweets> likedTweets = likesService.getLikedTweets(username);
-        List<TweetsDto> likedTweetsDtos;
+        List<TweetsDetailsDto> likedTweetsDtos;
         likedTweetsDtos = likedTweets.stream()
-                .map(tweetsMapper::mapTo)
+                .map(TweetsDetailsDtoMapper::mapTo)
                 .toList();
-        return likedTweetsDtos;
+        return ResponseEntity.ok(likedTweetsDtos);
     }
 
     // Get all users who liked a tweet
+    @Operation(summary = "Get users who liked a tweet")
     @GetMapping("/tweets/{tweetId}/likes/users")
-    public List<UserDto> getUsersWhoLikedTweet(@PathVariable int tweetId) {
-        List<UserDto> users;
+    public ResponseEntity<?> getUsersWhoLikedTweet(@PathVariable int tweetId) {
+        List<UserResponseDto> users;
         users = likesService.getUsersWhoLikedTweet(tweetId).stream()
-                .map(UserDto::new)
+                .map(userResponseMapper::mapTo)
                 .toList();
-        return users;
+
+        return ResponseEntity.ok(users);
     }
 }
